devutils/i18n: split generate/translate into separate libs
